Move isAvailable implementation
